Works, but ugly

+++ src/main/kotlin/com/rePUB/rePUB/model/User.kt
@@ -1,16 +1,14 @@
 package com.rePUB.rePUB.model
 
 import javax.persistence.Entity
-import javax.persistence.ForeignKey
 import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType
 import javax.persistence.Id
-import javax.validation.constraints.NotBlank
 
 @Entity
 data class User(
         @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-        val userId: Long = 0,
+        val userId: Int = 0,
 
         val name: String = ""
 )
